+++ FDI.Simple/AppIG4/OrderDetailAppIG4Item.cs
@@ -5,10 +5,10 @@ namespace FDI.Simple
 {
     public class OrderDetailAppIG4Item : BaseSimple
     {
-        public int ID { get; set; }
+        public long ID { get; set; }
         public int? OrderId { get; set; }
         public int? ProductId { get; set; }
-        public int? Quantity { get; set; }
+        public decimal? Quantity { get; set; }
         public string ProductName { get; set; }
         public string UrlPicture { get; set; }
         public string Address { get; set; }
